update main layout
