+++ Cargo.toml
@@ -30,6 +30,7 @@ ogg = "0.9.2"
 rand = "0.9.2"
 ratatui-image = "8.0.1"
 image = "0.25.8"
+libc = "0.2.176"
 [dependencies.windows-sys]
 version = ">=0.59, <=0.61"
 features = [
